cleanup - part II
reshuffle file structure to make more sense
